+++ src/modules/auction-house/CreateAuctionHouse.ts
@@ -64,42 +64,6 @@ export const createAuctionHouse = async (
     auctionHouse
   );
 
-  const rentExempt = await connection.getMinimumBalanceForRentExemption(0);
-  console.log('rentExempt', rentExempt);
-  const trBalance = await connection.getBalance(treasuryAccount);
-  console.log('treasuryAccount balance', trBalance);
-  const faBalance = await connection.getBalance(feeAccount);
-  console.log('feeAccount balance', faBalance);
-
-  if (trBalance < rentExempt || faBalance < rentExempt) {
-    const tx = new Transaction();
-    if (trBalance < rentExempt) {
-      const ix = SystemProgram.transfer({
-        fromPubkey: wallet.publicKey,
-        toPubkey: treasuryAccount,
-        lamports: rentExempt,
-      });
-      tx.add(ix);
-    }
-
-    if (faBalance < rentExempt) {
-      const ix = SystemProgram.transfer({
-        fromPubkey: wallet.publicKey,
-        toPubkey: feeAccount,
-        lamports: rentExempt,
-      });
-      tx.add(ix);
-    }
-
-    tx.feePayer = wallet.publicKey;
-    tx.recentBlockhash = (await connection.getLatestBlockhash()).blockhash;
-
-    const signed = await wallet.signTransaction(tx);
-
-    const txnId = await connection.sendRawTransaction(signed.serialize());
-    console.log(`rentExempt transfer Transaction ID: ${txnId}`);
-  }
-
   return createCreateAuctionHouseInstruction(
     {
       treasuryMint: tMintKey,
